+++ src/measure.js
@@ -1,7 +1,7 @@
 /*
  * The MIT License (MIT)
  *
- * Copyright (c) 2020 Karl STEIN
+ * Copyright (c) 2021 Karl STEIN
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -20,7 +20,6 @@
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
- *
  */
 
 import {
@@ -35,7 +34,7 @@ import {
  * @param {number} iterations
  * @return {*}
  */
-function measure(func, iterations = 1) {
+export function measureSync(func, iterations = 1) {
   const times = [];
 
   for (let i = 0; i < iterations; i += 1) {
@@ -61,5 +60,3 @@ export function logMeasureResult(result) {
     `slowest: ${formatMillis(result.slowest)} ms`,
   ].join('\r\n'));
 }
-
-export default measure;