+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v2.0.0
+
+- **[BREAKING CHANGE]** Renamed function `measure()` to `measureSync()`
+- **[BREAKING CHANGE]** Renamed function `benchmark()` to `benchmarkSync()`
+- Upgraded dependencies
+
 ## v1.1.5
 
 - Upgraded dependencies